results/rebuttal/run-20231111-034248_LLMatic-zero_shot_donnemartin-system-design-oop-finance_0_20-runs_log/finance/14/LLMatic/services.py
from models import User, Transaction, Category, BankAccount, Budget, Investment
from collections import defaultdict
from datetime import datetime
import hashlib


class UserService:
	def __init__(self):
		self.users = {}

	def create_user(self, username, password, email):
		if username in self.users:
			return 'User already exists'
		encrypted_password = hashlib.sha256(password.encode()).hexdigest()
		self.users[username] = User(username, encrypted_password, email)
		return 'User created successfully'

	def authenticate_user(self, username, password):
		encrypted_password = hashlib.sha256(password.encode()).hexdigest()
		if username in self.users and self.users[username].password == encrypted_password:
			return 'User authenticated'
		return 'Authentication failed'

	def recover_password(self, username):
		if username in self.users:
			return 'Password recovery link has been sent to your email'
		return 'User not found'

	def conduct_security_audit(self):
		# This is a placeholder for conducting security audits
		# In a real-world application, this would involve checking for any anomalies or suspicious activities
		return 'Security audit conducted successfully'

	def ensure_compliance(self):
		# This is a placeholder for ensuring compliance with financial regulations
		# In a real-world application, this would involve making sure that all financial transactions comply with the relevant regulations
		return 'Compliance check conducted successfully'

# Rest of the services remain the same

