results/rebuttal/run-20231111-034248_LLMatic-zero_shot_donnemartin-system-design-oop-finance_0_20-runs_log/finance/14/LLMatic/test_services.py
from services import UserService
from models import User
import hashlib

def test_user_service():
	user_service = UserService()
	assert user_service.create_user('test', 'test', 'test@test.com') == 'User created successfully'
	assert user_service.authenticate_user('test', 'test') == 'User authenticated'
	assert user_service.recover_password('test') == 'Password recovery link has been sent to your email'
	assert user_service.conduct_security_audit() == 'Security audit conducted successfully'
	assert user_service.ensure_compliance() == 'Compliance check conducted successfully'

# Rest of the tests remain the same


